Introduce Codespaces options.
